add Makefile and try_decrement()

+++ src/state.rs
@@ -11,7 +11,7 @@ pub static CONFIG_KEY: &[u8] = b"config";
 
 #[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, JsonSchema)]
 pub struct State {
-  pub count: i32,
+  pub count: u64,
   pub owner: Addr,
 }
